// lib/api/generated/schemas/adminAiUsageControllerListPeriod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * НеМалика API
 * Маркетплейс компьютерной техники — REST API для веб-витрины (Next.js) и Telegram mini-app. Покупать можно без авторизации; вход через Telegram нужен продавцу, администратору и покупателю, который хочет видеть историю просмотров на всех своих устройствах.
 * OpenAPI spec version: 1.0.0
 */

export type AdminAiUsageControllerListPeriod = typeof AdminAiUsageControllerListPeriod[keyof typeof AdminAiUsageControllerListPeriod];


export const AdminAiUsageControllerListPeriod = {
  all: 'all',
  today: 'today',
  '7d': '7d',
  '30d': '30d',
} as const;