// lib/api/generated/schemas/adminAiUsageControllerListSort.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * НеМалика API
 * Маркетплейс компьютерной техники — REST API для веб-витрины (Next.js) и Telegram mini-app. Покупать можно без авторизации; вход через Telegram нужен продавцу, администратору и покупателю, который хочет видеть историю просмотров на всех своих устройствах.
 * OpenAPI spec version: 1.0.0
 */

export type AdminAiUsageControllerListSort = typeof AdminAiUsageControllerListSort[keyof typeof AdminAiUsageControllerListSort];


export const AdminAiUsageControllerListSort = {
  newest: 'newest',
  oldest: 'oldest',
  cost_desc: 'cost_desc',
  cost_asc: 'cost_asc',
  credits_desc: 'credits_desc',
} as const;